Streamlit_App/requirements.txt
# pip install -r requirements.txt

streamlit>=0.82.0

#requirements for yolov5
# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML==5.4.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# logging -------------------------------------
tensorboard>=2.4.1
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.8.1

# scikit-learn==0.19.2  # for coreml quantization
thop  # FLOPS computation
pycocotools>=2.0  # COCO mAP


# requirements for cyclegan
dominate>=2.4.0
visdom>=0.1.8.8

#requirements for vgg16 feature extractor
tensorflow==2.4.1
